+++ src/group_report.ts
@@ -1,11 +1,10 @@
 import { InstanceTracker } from './instance_tracker';
 import { Context } from './context';
-import { InstanceGroup } from './instance_group';
 import { CloudInstance } from './cloud_manager';
 import ShutdownManager from './shutdown_manager';
 import MetricsLoop from './metrics_loop';
 import ReconfigureManager from './reconfigure_manager';
-import { InstanceState, JibriStatusState } from './instance_store';
+import { InstanceGroup, InstanceState, JibriStatusState } from './instance_store';
 
 export interface InstanceReport {
     instanceId: string;
