+++ src/metrics_loop.ts
@@ -1,10 +1,10 @@
 import { Redis } from 'ioredis';
 import * as promClient from 'prom-client';
-import InstanceGroupManager, { InstanceGroup } from './instance_group';
+import InstanceGroupManager from './instance_group';
 import { Context } from './context';
 import { InstanceTracker } from './instance_tracker';
 import { CloudInstance } from './cloud_manager';
-import { InstanceState } from './instance_store';
+import { InstanceGroup, InstanceState } from './instance_store';
 
 const groupsManaged = new promClient.Gauge({
     name: 'autoscaling_groups_managed',
