+++ src/oracle_instance_manager.ts
@@ -1,12 +1,12 @@
 import core = require('oci-core');
 import common = require('oci-common');
 import identity = require('oci-identity');
-import { InstanceGroup } from './instance_group';
 import { Context } from './context';
 import { ResourceSearchClient } from 'oci-resourcesearch';
 import * as resourceSearch from 'oci-resourcesearch';
 import { CloudRetryStrategy } from './cloud_manager';
 import { AbstractCloudInstanceManager, CloudInstanceManager, CloudInstance } from './cloud_instance_manager';
+import { InstanceGroup } from './instance_store';
 
 interface FaultDomainMap {
     [key: string]: string[];
